changed naming convention attempt on fixing code style
