--- src/main/java/seedu/address/model/camp_appointment/Participant.java
@@ -1,4 +1,0 @@
-package seedu.address.model.camp_appointment;
-
-public class Participant {
-}
