+++ src/main/java/seedu/address/model/role/Participant.java
@@ -1,0 +1,7 @@
+package seedu.address.model.role;
+
+/**
+ * Participant for FOP.
+ */
+public class Participant {
+}
